+++ components/ui/Header.vue
@@ -1,7 +1,9 @@
 <template>
 	<header class="header">
 		<div class="container-fluid flex">
-			<h1 class="header-title"><nuxt-link to="/">Parkour<span>Finder</span></nuxt-link></h1>
+			<h1 class="header-title">
+				<nuxt-link to="/">Parkour<span>Finder</span></nuxt-link>
+			</h1>
 
 			<ui-menu/>
 		</div>
@@ -26,14 +28,14 @@ export default {
 
 .header
 	background $color-white
-	padding .5rem 0
+	padding 0.5rem 0
 	box-shadow 0 0 5px $color-shadow
 	position relative
 	z-index 1000
 
 .header-title
 	margin 0
-	font 400 $font-size-heading-3/1 $font-stack-heading
+	font 400 $font-size-heading-3 / 1 $font-stack-heading
 	color $color-red
 	transform rotate(-8deg)
 	white-space nowrap
@@ -41,17 +43,21 @@ export default {
 	@media $mq-tablet
 		font-size $font-size-title
 
+
 	a
 		color inherit
 		text-decoration none
-		transition .25s ease
+		transition 0.25s ease
 
 		&:hover
 			color darken($color-red, 15%)
 
+
+
 	span
 		display inline-block
-		transform translateY(.225em)
+		transform translateY(0.225em)
+
 
 .header-actions
 	margin-left auto