+++ store/map.js
@@ -3,13 +3,13 @@ export const state = () => ({
 })
 
 export const mutations = {
-	setMarkers(state, markers) {
+	setMarkers (state, markers) {
 		state.markers = markers
 	}
 }
 
 export const actions = {
-	async init({ commit }) {
+	async init ({ commit }) {
 		let { data } = await this.$axios.get(process.env.api.spots)
 		commit('setMarkers', data)
 	}