+++ store/spot.js
@@ -1,0 +1,9 @@
+export const state = () => ({
+	currentSpot: null
+})
+
+export const mutations = {
+	setCurrentSpot (state, spot) {
+		state.currentSpot = spot
+	}
+}